requirements.txt
appdirs==1.4.4
asdl==0.1.5
astor==0.8.1
attrs==18.2.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
corenlp-protobuf==3.8.0
decorator==4.4.2
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
jieba==0.42.1
joblib==0.17.0
jsonnet==0.16.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.6.0
networkx==2.4
nltk==3.5
numpy==1.19.1
olefile==0.46
packaging==20.4
pathspec==0.8.0
pglast==1.12
Pillow @ file:///tmp/build/80754af9/pillow_1603822253009/work
pluggy==0.13.1
protobuf==3.13.0
psutil==5.7.3
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.1
regex==2020.10.28
requests==2.24.0
sacremoses==0.0.43
sentencepiece==0.1.94
six==1.15.0
stanford-corenlp @ git+https://github.com/stanfordnlp/python-stanford-corenlp.git@57a739e3aa97942e443bea63bd413afcbe5e7f3f
stanfordcorenlp==3.9.1.1
stanza==1.0.1
timeout-decorator==0.4.1
tokenizers==0.9.2
toml==0.10.2
torch==1.6.0
torchtext==0.7.0
torchvision==0.7.0
tqdm==4.48.2
transformers==3.4.0
typed-ast==1.4.1
urllib3==1.25.11
zipp==3.4.0
tensorboard==2.4